requirements.txt
Django==1.7.10
Markdown==2.6.5
Pillow==3.1.1
defusedxml==0.4.1
django-allauth==0.24.1
django-bootstrap-breadcrumbs==0.7.3
django-crispy-forms==1.5.2
django-dajax==0.9.2
django-dajaxice==0.7
django-grappelli==2.6.5
django-markdown==0.8.4
django-modalview==0.1.5
django-wkhtmltopdf==3.0.0
oauthlib==1.0.3
psycopg2==2.6.1
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.1
